feat: run rv64imac alongside rv32
M17, the RV64 transition: the core stops being RV32-only and runs
RV64IMAC too, selected per program from the ELF class rather than by
a separate build. XLEN is a runtime property (cpu->xlen): every
register, the PC, the CSRs, and every address are stored as 64 bits,
and a Spike-style sext_xlen invariant (in RV32 a register holds the
sign-extension of its 32-bit value) keeps the executor mostly
width-agnostic -- add/compare/logic are right once reg_write
re-sign-extends; only the shifts, the *W ops, and the loads carry an
explicit width branch. The MMU masks a VA to XLEN at its single
choke point, and RV64 runs Bare (Sv32 is an RV32 scheme; Sv39 is M18).

Adds RV64I (*W ops, LD/SD/LWU, 6-bit shifts), RV64M (a portable
128-bit mulh) and RV64A (.D doublewords), and RV64C by the same
expand-to-32-bit path (C.ADDIW/C.LD/C.SD/C.LDSP/C.SDSP/C.SUBW/C.ADDW).
The CSRs become XLEN-wide and the mcause interrupt bit moves to bit
63; the ELF64 loader, disassembler, GDB stub (riscv:rv64, 64-bit
register packets), and boot device tree are all width-aware. Every
RV64-only encoding traps illegal in RV32, so the RV32 suite stays
bit-for-bit green -- the refactor's safety checkpoint. RV32F/D stay
deferred (M11), so this is RV64IMAC, not the full GC.

New tests/rv64/ conformance suite and make check-rv64: quanta exit-0
plus a qemu-riscv64 differential on the user-mode programs (the
golden-model net). Verified clean under ASan/UBSan and the
cppcheck/clang-tidy static analysis; version bumped to 0.2.0.

+++ src/device.c
@@ -12,7 +12,7 @@
 
 /* True when `addr` lies in [base, base+size). Written to stay correct for
  * addresses below base (the unsigned difference wraps high, failing the test). */
-static int in_window(uint32_t addr, uint32_t base, uint32_t size) {
+static int in_window(uint64_t addr, uint32_t base, uint32_t size) {
     return (addr - base) < size;
 }
 
@@ -21,7 +21,7 @@ void plat_init(Platform *p) {
     p->clint.mtimecmp = (uint64_t)-1; /* parked: no timer interrupt until armed */
 }
 
-int plat_contains(uint32_t addr) {
+int plat_contains(uint64_t addr) {
     return in_window(addr, CLINT_BASE, CLINT_SIZE) ||
            in_window(addr, PLIC_BASE,  PLIC_SIZE)  ||
            in_window(addr, UART_BASE,  UART_SIZE);
@@ -188,7 +188,7 @@ static void clint_write(Clint *c, uint32_t off, uint32_t val) {
  * MMIO dispatch and interrupt pull.
  * ------------------------------------------------------------------------ */
 
-uint32_t plat_read(Platform *p, uint32_t addr, uint32_t size) {
+uint32_t plat_read(Platform *p, uint64_t addr, uint32_t size) {
     if (in_window(addr, UART_BASE, UART_SIZE)) /* byte registers */
         return uart_read(&p->uart, addr - UART_BASE);
 
@@ -207,7 +207,7 @@ uint32_t plat_read(Platform *p, uint32_t addr, uint32_t size) {
     return word;
 }
 
-void plat_write(Platform *p, uint32_t addr, uint32_t size, uint32_t value) {
+void plat_write(Platform *p, uint64_t addr, uint32_t size, uint32_t value) {
     (void)size; /* CLINT/PLIC take word writes; the UART is byte-addressed */
     if (in_window(addr, UART_BASE, UART_SIZE))
         uart_write(&p->uart, addr - UART_BASE, (uint8_t)value);
